requirements.txt
python-dotenv
gspread
PyDrive
dataset

google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib